Fixing up the Electron app so that it works
