--- front/main.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Electron Test App</title>
-    </style>
-  </head>
-  <body>
-    <div id="screenPublisher"></div>
-    <video id="screenPublisherVideo"></video>
-    <script src="/js/commons.min.js" type="text/javascript" charset="utf-8"></script>
-    <script src="/js/app.bundle.min.js" type="text/javascript" charset="utf-8"></script>
-    <script src="file:///Users/James/OPENTOK/electron-screenshare-opentok/app/public/js/opentok.js" charset="utf-8"></script>
-  </body>
-</html>
