--- front/main.js
@@ -1,21 +1,0 @@
-var OT_API_KEY = "100";
-var OT_SESSION_ID = "1_MX4xMDB-fjE0OTc0NjU2MjU5MjZ-bGtWdUtMaFpCU1BEbUJ5TG9kOXYwc3Myfn4";
-var OT_TOKEN = "T1==cGFydG5lcl9pZD0xMDAmc2RrX3ZlcnNpb249dGJwaHAtdjAuOTEuMjAxMS0wNy0wNSZzaWc9NTNlMzQwYzhmMjU5MTYwOTU4ZDhlOTYxOGYyNDMxOWUxNTZhYmFmYjpzZXNzaW9uX2lkPTFfTVg0eE1EQi1makUwT1RjME5qVTJNalU1TWpaLWJHdFdkVXRNYUZwQ1UxQkViVUo1VEc5a09YWXdjM015Zm40JmNyZWF0ZV90aW1lPTE0OTc0NjU2MjYmcm9sZT1tb2RlcmF0b3Imbm9uY2U9MTQ5NzQ2NTYyNi4wNjk0MzYzNzAyODYwJmV4cGlyZV90aW1lPTE1MDAwNTc2MjY=";
-
-window.addEventListener('load', function () {
-
-  const session = OT.initSession(OT_API_KEY, OT_SESSION_ID);
-  session.on('streamCreated', function (event) {
-    session.subscribe(event.stream, function (err) {
-      if (err) alert(err.message);
-    });
-  });
-
-  var publisher = OT.initPublisher("myPublisher", function(err) {
-    if (err)
-      console.log('OT.initPublisher error: ', err);
-    else
-      console.log('Publisher Initialized');
-    });
-});
-
