+++ index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Electron Test App</title>
+    <script src="https://tbdev.tokbox.com/v2/js/opentok.min.js" charset="utf-8"></script>
+    </style>
+  </head>
+  <body>
+    <script src="app.js" charset="utf-8"></script>
+  </body>
+</html>
