+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "opentok-electron",
   "version": "1.0.0",
-  "description": "a opentok electron hello world sample app",
-  "main": "app.js",
+  "description": "an opentok electron hello world sample app",
+  "main": "main.js",
   "scripts": {
-    "test": "npm test"
+    "start": "electron ."
   },
   "repository": {
     "type": "git",
@@ -20,15 +20,9 @@
     "url": "https://github.com/zjb421/opentok-electron/issues"
   },
   "homepage": "https://github.com/zjb421/opentok-electron#readme",
-  "bin": {
-    "my-app": "./bin/my-app"
-  },
-  "devDependencies": {
-    "electron": "^1.3.4"
-  },
   "dependencies": {
     "electron-prebuilt": "^1.4.13",
     "electron-rebuild": "^1.5.11",
-    "express": "^4.15.3"
+    "electron": "^1.3.4"
   }
 }
